Move ConfigActions into actions folder
